Add function to count number of unpacked varying vectors
This counts the number of vectors used by each active varying in a
shader. This can be used to ensure that a shader will succesfully
compile without using more than GL_MAX_VARYING_VECTORS on devices that
do not perform spec-compliant packing of varyings. This is in contrast
to sh::CheckVariablesWithinPackingLimits(), which checks whether there
are enough vectors *with* spec-compliant packing.